Adds a workflow to npm publish when a tag is pushed.
You'll need to add an `NPM_TOKEN` secret to this repository. In this repos settings, there's a nav item on the left called "secrets". In your account settings in npm you can generate a token to use as a secret here.

Once entered, the GitHub secret won't be seen, even by collaborators.

I'm 90% sure this'll work, but we won't know until the secret is added and a tag is pushed!